code/requirements2b.txt
flask_migrate==2.5.3
flask_sqlalchemy==2.4.3
flask_wtf==0.14.3
flask_login==0.5.0
fhirclient==3.2.0
